Fixed windows warning.
svn path=/trunk/libavg/; revision=4250

+++ src/anim/SimpleAnim.cpp
@@ -113,7 +113,7 @@ bool SimpleAnim::step()
     }
 }
 
-double SimpleAnim::getStartTime() const
+long long SimpleAnim::getStartTime() const
 {
     return m_StartTime;
 }
